+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-BHN3G-iF.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-DGiUgnlp.js";import{_ as a}from"./mermaid.core-ehl9LlA0.js";import"./chunk-5VM5RSS4-C92AZ4Vc.js";import"./chunk-XXDRQBXY-CsDludLE.js";import"./chunk-POPQ4Y6H-CN1bYFZP.js";import"./chunk-F27PBJKO-CksW1kwI.js";import"./channel--PyrDM_k.js";import"./index-DQJvMNyc.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};